# triedb/pathdb:  use binary.append to eliminate the tmp scratch slice