`binary.AppendUvarint` offers better performance than using append directly, because it avoids unnecessary memory allocation and copying. 

In our case, it can increase the performance by +35.8% for the `blockWriter.append` function:

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkBlockWriterAppend-8     5.97          3.83          -35.80%
```